move cmetamodel to legacy
