+++ legacy/metamodel/nn_generator/nn_settings.py
